created exceptions and test structure

+++ squadmap-backend/src/main/java/de/adesso/squadmap/service/employee/DeleteEmployeeService.java
@@ -2,6 +2,7 @@ package de.adesso.squadmap.service.employee;
 
 import de.adesso.squadmap.port.driver.employee.delete.DeleteEmployeeUseCase;
 import de.adesso.squadmap.repository.EmployeeRepository;
+import de.adesso.squadmap.exceptions.EmployeeNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,6 +16,9 @@ public class DeleteEmployeeService implements DeleteEmployeeUseCase {
 
     @Override
     public void deleteEmployee(Long employeeId) {
+        if(!employeeRepository.existsById(employeeId)){
+            throw new EmployeeNotFoundException();
+        }
         employeeRepository.deleteById(employeeId);
     }
 }
