+++ squadmap-backend/src/main/java/de/adesso/squadmap/service/project/DeleteProjectService.java
@@ -2,6 +2,7 @@ package de.adesso.squadmap.service.project;
 
 import de.adesso.squadmap.port.driver.project.delete.DeleteProjectUseCase;
 import de.adesso.squadmap.repository.ProjectRepository;
+import de.adesso.squadmap.exceptions.ProjectNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,7 +15,10 @@ public class DeleteProjectService implements DeleteProjectUseCase {
     }
 
     @Override
-    public void deleteProject(Long projectId) {
+    public void deleteProject(Long projectId){
+        if (!projectRepository.existsById(projectId)){
+            throw new ProjectNotFoundException();
+        }
         projectRepository.deleteById(projectId);
     }
 }
